creates room form
