+++ app/dashboard/page.tsx
@@ -1,0 +1,12 @@
+import RoomForm from "../Components/Forms/RoomForm"
+
+export default function Dashboard() {
+    return (
+        <>
+            <div>
+                <h2>Post a room: </h2>
+                <RoomForm />        
+            </div>
+        </>
+    )
+}
